api/pyproject.toml
[tool.poetry]
name = "ai-sdk-preview-rag"
version = "0.1.0"
description = ""
authors = ["Nick Yu <wiseyu@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">3.12,<3.13"
supabase = "^2.10.0"
inngest = "^0.4.16"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
python-dotenv = "^1.0.1"
pydantic-settings = "^2.6.1"
litellm = "^1.52.6"
html2text = "^2024.2.26"
markdown-it-py = "^3.0.0"
flex-swarm = "^0.1.1"
gotrue = "^2.11.0"
fastapi = "^0.115.6"


[tool.poetry.group.dev.dependencies]
ollama = "^0.4.4"
playwright = "^1.49.1"
nest-asyncio = "^1.6.0"
fastapi-cli = "^0.0.7"
undetected-playwright = "^0.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
